Don't allow another thread (other than the one doing the reading) to change a block out of state READING.
This simplifies the logic in the block cache when the kernel attempts to
write the same block that is currently being read by another thread. This
case probably never happens, but if it does, instead of allowing both
operations to proceed simultaneously we will now block the writer thread
until the reader thread finishes. This change is needed to facilitate
other upcoming changes.

This also fixes an obscure corner case where two threads could end up
reading the same block at the same time if there were an intervening
write.