Fix use-after-free
I wrote this on Fri, 25 Apr 2014 against s3backer 1.3.7 when doing
consulting for SoftNAS. The details of this use-after-free have been
lost to time, but my recollection is that they had a crash where another
thread had the old pointer on the stack after realloc() had assigned a
new pointer. The solution was deleting this code.

I maintained copyright ownership of the "change" (really just deleting
code) as well as all other changes that I made for them. I am fully able
to submit it under the GPL. Back then, getting in touch with the project
maintainer was hard, but since I see it is on github, it is time to send
the patch.

Signed-off-by: Richard Yao <ryao@gentoo.org>