Added the Delete Multiple Object support.
It works by intercepting every immediate block deletion request in the http_io_write_block function, and assigning it to a secondary separate thread which is responsible of collecting all these requests for at most 25ms (or up to 1000 objects). When timeout occurs or the list is full this secondary function fires an http_io POST request with all the collected keys, resulting in one single request to S3.