Feature: Add support for ONEZONE_IA
This enables support for users to directly configure `ONEZONE_IA` as a
target storage class for files placed into S3.  This is important
because when using LifeCycle policies users are forced to keep content
in a higher storage class for a minimum of 30 days.

When specified at the actual client level, content can be immediately
specified as the preferred storage class.