Fix ETag handling bug triggered by server side encryption (issue #135).
In the early days of S3 when s3backer was written, the ETag for a block
was always the same as the block's MD5 hash value. With the advent of
server-side encryption (SSE), this is no longer true.

As a result, when SSE is used s3backer can generate phantom I/O errors
caused by a mismatch between the block's ETag as recorded in the MD5
cache vs. its actual ETag.

The solution is simply to always report back the ETag explicitly provided
in the HTTP PUT response rather than conflating it with the MD5 checksum.

We still require however that the returned ETag look like an MD5 checksum,
i.e., 32 hex digits. This should always be the case as long as no partial
object updates are performed.