Reduce redundant page cache entries when using disk-based block cache.
Currently, there will often be two copies of the same data in the page cache: one for the
data in the FUSE filesystem, and one for the (same) data in the block cache file.

This patch adds explicit posix_fadvise() hints to the kernel that the page cache entries
for the block cache file are not needed.

It is not clear if this will work when huge pages are in use, but will do no harm in this
case.

It would be preferable if there was a way to flag the entire file as not needing to be
cached, but unfortunately this isn't possible (O_DIRECT comes close, but depends on
support from the underlying filesystem and requires writes and reads to be page-aligned).