Fix data corruption when kernel I/O block size != s3backer block size (#184).
When these block sizes are different, it's possible for the kernel to be
reading and/or writing (different parts of) the same block simultaneously.
Because partial block writes are performed by reading, patching, then writing
back the entire block, this means data writes can get lost unless we impose
read/write locking for each block.

As part of this change, simplify the code by requiring all partial read/write
operations to be done at the top layer of the stack (i.e., FUSE or NBD layer).
This eliminates the "read_block_part" and "write_block_part" layer functions.