Recalculate date and authorization on retry (issue #214).
Previously we were adding the date header and authorization headers
before each request attempt, which means they didn't change if the
request is retried. Now we move the addition of those headers into
the "prepper" functions so they are recalculated on each attempt.
Of course this means we also have to find and remove any old ones.